moved flagset and flagset filter classes to storage
